refactor: move some more tests
